create_app function and test upgrade
create_app function integrated into init app which creates an
flask app instance and initialize the app using the config file
passed in as argument. The function intializes the db, migrates to
the latest version and login form. Finally a blueprint is registered
before the app instance is returned.

The function is used within test.py to call a app instance with
a non-persistent db for testing